// src/generated-client/models/BeansGet200ResponseBeansInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Coffee Addiction API
 * Coffee Addiction API swagger documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BeansGet200ResponseBeansInner
 */
export interface BeansGet200ResponseBeansInner {
    /**
     * Unique identifier for the coffee bean
     * @type {number}
     * @memberof BeansGet200ResponseBeansInner
     */
    beanId?: number;
    /**
     * Name of the coffee bean
     * @type {string}
     * @memberof BeansGet200ResponseBeansInner
     */
    name?: string;
    /**
     * Country or region of origin
     * @type {string}
     * @memberof BeansGet200ResponseBeansInner
     */
    origin?: string;
    /**
     * Level of roast (e.g., light, medium, dark)
     * @type {string}
     * @memberof BeansGet200ResponseBeansInner
     */
    roastLevel?: string;
    /**
     * URL
     * @type {string}
     * @memberof BeansGet200ResponseBeansInner
     */
    imageUrl?: string;
    /**
     * Price per kilogram of coffee beans
     * @type {number}
     * @memberof BeansGet200ResponseBeansInner
     */
    pricePerKg?: number;
    /**
     * Available quantity in stock
     * @type {number}
     * @memberof BeansGet200ResponseBeansInner
     */
    stockQuantity?: number;
    /**
     * Detailed description of the coffee bean
     * @type {string}
     * @memberof BeansGet200ResponseBeansInner
     */
    description?: string;
}

/**
 * Check if a given object implements the BeansGet200ResponseBeansInner interface.
 */
export function instanceOfBeansGet200ResponseBeansInner(value: object): value is BeansGet200ResponseBeansInner {
    return true;
}

export function BeansGet200ResponseBeansInnerFromJSON(json: any): BeansGet200ResponseBeansInner {
    return BeansGet200ResponseBeansInnerFromJSONTyped(json, false);
}

export function BeansGet200ResponseBeansInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): BeansGet200ResponseBeansInner {
    if (json == null) {
        return json;
    }
    return {
        
        'beanId': json['beanId'] == null ? undefined : json['beanId'],
        'name': json['name'] == null ? undefined : json['name'],
        'origin': json['origin'] == null ? undefined : json['origin'],
        'roastLevel': json['roastLevel'] == null ? undefined : json['roastLevel'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'pricePerKg': json['pricePerKg'] == null ? undefined : json['pricePerKg'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function BeansGet200ResponseBeansInnerToJSON(json: any): BeansGet200ResponseBeansInner {
    return BeansGet200ResponseBeansInnerToJSONTyped(json, false);
}

export function BeansGet200ResponseBeansInnerToJSONTyped(value?: BeansGet200ResponseBeansInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'beanId': value['beanId'],
        'name': value['name'],
        'origin': value['origin'],
        'roastLevel': value['roastLevel'],
        'imageUrl': value['imageUrl'],
        'pricePerKg': value['pricePerKg'],
        'stockQuantity': value['stockQuantity'],
        'description': value['description'],
    };
}